Working calculator

+++ src/main/java/com/breadsb/simpleapp/Calculations.java
@@ -2,8 +2,8 @@ package com.breadsb.simpleapp;
 
 public class Calculations {
 
-    public int calculate_EEI(double sprawnoscNominalna, double wspolczynnikBiomasy, double parametrF2) {
-        return (int) ((sprawnoscNominalna * wspolczynnikBiomasy)
+    public double calculate_EEI(double sprawnoscNominalna, double wspolczynnikBiomasy, double parametrF2) {
+        return ((sprawnoscNominalna * wspolczynnikBiomasy)
                 - ( 0.1*( sprawnoscNominalna*wspolczynnikBiomasy ) )
                 + parametrF2);
     }
